Turned Motion Sensor into an upgrade. Closes #2241.
