+++ src/main/scala/li/cil/oc/integration/opencomputers/DriverMotionSensor.scala
@@ -1,0 +1,31 @@
+package li.cil.oc.integration.opencomputers
+
+import li.cil.oc.Constants
+import li.cil.oc.api
+import li.cil.oc.api.driver.EnvironmentProvider
+import li.cil.oc.api.driver.item.HostAware
+import li.cil.oc.api.network.EnvironmentHost
+import li.cil.oc.common.Slot
+import li.cil.oc.common.Tier
+import li.cil.oc.server.component
+import net.minecraft.item.ItemStack
+
+object DriverMotionSensor extends Item with HostAware {
+  override def worksWith(stack: ItemStack) = isOneOf(stack,
+    api.Items.get(Constants.BlockName.MotionSensor))
+
+  override def createEnvironment(stack: ItemStack, host: EnvironmentHost) =
+    if (host.world != null && host.world.isRemote) null
+    else new component.MotionSensor(host)
+
+  override def slot(stack: ItemStack) = Slot.Upgrade
+
+  override def tier(stack: ItemStack) = Tier.Two
+
+  object Provider extends EnvironmentProvider {
+    override def getEnvironment(stack: ItemStack): Class[_] =
+      if (worksWith(stack))
+        classOf[component.MotionSensor]
+      else null
+  }
+}
